LPS-146050 Move all files related to ColorPicker to a new folder
