+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/ColorPicker/parseColorValue.js
@@ -12,8 +12,8 @@
  * details.
  */
 
-import {convertRGBtoHex} from './convertRGBtoHex';
-import {getValidHexColor} from './getValidHexColor';
+import {convertRGBtoHex} from '../../../app/utils/convertRGBtoHex';
+import {getValidHexColor} from '../../../app/utils/getValidHexColor';
 
 const ERROR_MESSAGES = {
 	mutuallyReferenced: Liferay.Language.get(
