+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/common/components/ColorPicker/DropdownColorPicker.test.js
@@ -21,7 +21,7 @@ import {
 } from '@testing-library/react';
 import React from 'react';
 
-import {DropdownColorPicker} from '../../../../src/main/resources/META-INF/resources/page_editor/common/components/DropdownColorPicker';
+import {DropdownColorPicker} from '../../../../../src/main/resources/META-INF/resources/page_editor/common/components/ColorPicker/DropdownColorPicker';
 
 const CONFIG = {
 	tokenReuseEnabled: true,
